bringing getFileHeader to PlainTextOutputGenerator to remove duplicated code
