+++ lib/src/presentation/output_print_generator/markdown_output_generator.dart
@@ -1,11 +1,12 @@
 import 'package:pull_request_coverage/src/domain/analyzer/models/analysis_result.dart';
-import 'package:pull_request_coverage/src/domain/input_reader/diff_reader/models/file_diff.dart';
 import 'package:pull_request_coverage/src/domain/user_options/models/markdown_mode.dart';
 import 'package:pull_request_coverage/src/domain/user_options/models/user_options.dart';
 import 'package:pull_request_coverage/src/presentation/output_print_generator/get_result_table.dart';
 import 'package:pull_request_coverage/src/presentation/output_print_generator/plain_text_output_generator.dart';
+import 'package:pull_request_coverage/src/presentation/use_case/colorize_cli_text.dart';
 
 class MarkdownOutputGenerator with PlainTextOutputGenerator {
+  @override
   final UserOptions userOptions;
   final GetResultTable getResultTable;
   const MarkdownOutputGenerator({
@@ -32,21 +33,7 @@ class MarkdownOutputGenerator with PlainTextOutputGenerator {
   }
 
   @override
-  String? getFileHeader(FileDiff fileDiff) {
-    final filePath = fileDiff.path;
-    final uncoveredLinesCount = fileDiff.uncoveredNewLinesCount;
-    final totalNewLinesCount = fileDiff.newLinesCount;
-    final ignoredMsg = fileDiff.ignoredUntestedLinesCount > 0
-        ? " / ${"${fileDiff.ignoredUntestedLinesCount} untested and ignored"}"
-        : "";
-    if (uncoveredLinesCount == 0) {
-      if (userOptions.reportFullyCoveredFiles) {
-        return " - `$filePath` is fully covered (${"+$totalNewLinesCount"}$ignoredMsg)";
-      }
-      return null;
-    }
-    return " - `$filePath` has $uncoveredLinesCount uncovered lines (${"+$totalNewLinesCount"}$ignoredMsg)\n";
-  }
+  String formatFileHeader(String text) => " - `$text";
 
   @override
   String? getLine(String line, int lineNumber, bool isANewLine, bool isUntested) {
@@ -72,4 +59,7 @@ class MarkdownOutputGenerator with PlainTextOutputGenerator {
     }
     return "\n${getResultTable(analysisResult)}";
   }
+
+  @override
+  ColorizeCliText? get colorizeCliText => null;
 }
